# src/main/java/com/ulises/posverse/rest/api/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ulises.posverse.rest.api;

import com.ulises.posverse.rest.api.dto.requests.ProductCreationRequestDTO;
import com.ulises.posverse.rest.api.dto.responses.ProductCreationResponseDTO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-07-17T23:45:41.592781600+02:00[Europe/Madrid]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "products", description = "the products API")
public interface ProductsApi {

    @Operation(
            operationId = "createProduct",
            summary = "Creates a product",
            responses = {
                    @ApiResponse(responseCode = "201", description = "Created", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ProductCreationResponseDTO.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/products",
            produces = { "application/json" },
            consumes = { "application/json" }
    )

    ResponseEntity<ProductCreationResponseDTO> createProduct(
            @Parameter(name = "ProductCreationRequestDTO", description = "Product creation request", required = true)
            @Valid @RequestBody final ProductCreationRequestDTO productCreationRequestDTO
    );

}